// models/interfaces.ts
export interface Produto {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: Rating;
}
export interface Rating {
    rate: number;
    count: number;
}
export interface User {
    id: number;
    name: string;
    email: string;
    username: string;
}